+++ src/main/java/com/example/Dingle/crime/controller/CrimeController.java
@@ -1,6 +1,6 @@
-package com.example.Dingle.environment.controller;
+package com.example.Dingle.crime.controller;
 
-import com.example.Dingle.environment.service.EnvironmentService;
+import com.example.Dingle.crime.service.CrimeService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -9,9 +9,9 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/v1/environment")
-public class EnvironmentController {
+public class CrimeController {
 
-    private final EnvironmentService environmentService;
+    private final CrimeService environmentService;
 
     @PostMapping("/crime")
     public void saveCctvInfra() throws Exception {
