+++ src/main/java/com/example/Dingle/crime/repository/CrimeProneAreaRepository.java
@@ -1,6 +1,6 @@
-package com.example.Dingle.environment.repository;
+package com.example.Dingle.crime.repository;
 
-import com.example.Dingle.environment.entity.CrimeProneArea;
+import com.example.Dingle.crime.entity.CrimeProneArea;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
